Feat default admin (#2)
* installed class validator,class-transformer

* created user entity

* created user-dto

* user entity modified

* user dto modified

* configured database with user entity

* Db setup

* installed nestjs/config

* env variables loaded

* env variable modified

* getALLusers created

* role added

* dotenv configured

* create ,delet,getAll  user features added

* deleted file

* removed db.config

* modified script file

* env example file

* config module modified

* deleted app .config redundant ENV_MODE

* removed import app.module

* decorated email with ISEmail

* removed SYNC and AUTOLOAD use as hardcoded

* default role modified

* create default admin feature added

* bcrypt downloaded

* class validation options selected

* default admin with checks

* role validation and bcrypt passwords

* role checks

* manual admin creation check

* protect admin deletion

* modified

* enums used

* enum created

* corrected delete logic

* removed file

* modified enum implementation

+++ src/user/user.service.ts
@@ -3,7 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './user.entity';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dtos/create-user.dto';
-
+import { userRole } from './enums/role.enum';
 @Injectable()
 export class UserService {
 
@@ -29,7 +29,11 @@ export class UserService {
     public async createUser(userDto: CreateUserDto) {
         try {
 
-            //          check is user with same username/gmail exist
+            if(userDto.role === userRole.ADMIN ){
+                throw new BadRequestException('Cannot create manual ADMIN')
+            }
+
+            //          check is user with same username/email exist
             const existingUser = await this.userRepository.findOne({
                 where: [{ username: userDto.username }, { email: userDto.email }]
             })
@@ -37,9 +41,12 @@ export class UserService {
                 throw new BadRequestException(`There is already a user with same email/username`)
 
             }
-
+            
             //          create user
-            let user = this.userRepository.create(userDto)
+            let user = this.userRepository.create({
+                ...userDto,
+                role:userRole.MEMBER
+            })
 
             //          save user
             await this.userRepository.save(user)
@@ -47,8 +54,6 @@ export class UserService {
             return {
                 message: "User created succesfully!!!"
             }
-
-
 
         } catch (error) {
             if (error.code === 'ECONNREFUSED') {
@@ -67,11 +72,15 @@ export class UserService {
         if (!user) {
             throw new NotFoundException("no user found")
         }
+        if(user.role === userRole.ADMIN ){
+            throw new BadRequestException('Cannot delete default ADMIN')
+        }
 
         await this.userRepository.delete(id)
         return {
             message: `user with id ${id} deleted successfully`
         }
     }
+
 
 }
